% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19-package.R
\docType{package}
\name{covid19-package}
\alias{covid19}
\alias{covid19-package}
\title{covid19: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Cuiping Wei \email{cwei0018@student.monash.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
