% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch shiny app}
\usage{
launch_app()
}
\value{
Shiny app -- COVID-19 Global Cases
}
\description{
This function can be used to launch shiny app in package.
}
\examples{
\dontrun{
library(covid19)
launch_app()
}


}
