% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_highchart.R
\name{plot_highchart}
\alias{plot_highchart}
\title{Title}
\usage{
plot_highchart(df, x, y, z, ylabs, xlabs, title)
}
\arguments{
\item{df}{The data}

\item{x}{The date of COVID19 cases}

\item{y}{The varible you interested}

\item{ylabs}{y label}

\item{xlabs}{x label}

\item{title}{plot title}
}
\value{
plot
}
\description{
This function is to produce highchart plot
}
