% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_linechart.R
\name{plot_linechart}
\alias{plot_linechart}
\title{Linechart}
\usage{
plot_linechart(df, var)
}
\arguments{
\item{df}{The data used in producing linechart.}

\item{var}{A character vector.}
}
\value{
Linechart
}
\description{
This function creates a linechart.
}
\examples{
library(magrittr)
library(coronavirus)
library(ggplot2)

data("coronavirus")
data <- coronavirus \%>\%
 dplyr::filter(date >= '2020-10-10') \%>\%
 dplyr::filter(country \%in\% c("France", "India", "Austria")) \%>\%
 tidyr::pivot_wider(names_from = type, values_from = cases)

 plot_linechart(data, "India")



}
