% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_data.R
\docType{data}
\name{pop_data}
\alias{pop_data}
\title{2019 Population data}
\format{
a tibble with 221 observations and 2 variables:
\itemize{
\item \strong{country}: Country name
\item \strong{population}: Population data for 2019
}
}
\source{
\href{https://databank.worldbank.org/home.aspx}{World Bank}
}
\usage{
pop_data
}
\description{
A dataset containing 2019 population data for 221 countries.
}
\examples{
pop_data


}
\keyword{datasets}
