% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valuebox.R
\name{valuebox}
\alias{valuebox}
\title{Shiny ValueBox output}
\usage{
valuebox(ID, width)
}
\arguments{
\item{ID}{The ID of valueBox.}

\item{width}{The width of valueBox.}
}
\value{
ValueBox
}
\description{
Shiny ValueBox output
}
