% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_type.R
\name{input_type}
\alias{input_type}
\title{Title}
\usage{
input_type(type, df = NULL, df1 = NULL, ID, label)
}
\arguments{
\item{type}{Iuput type}

\item{df}{first data}

\item{df1}{second data}

\item{ID}{ID}

\item{label}{label}
}
\value{

}
\description{
Title
}
