% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.R
\name{page}
\alias{page}
\title{Shiny dashboard page}
\usage{
page(title, body)
}
\arguments{
\item{title}{The title of dashboard page}

\item{body}{The dashboard body content}
}
\value{
Shiny dashboard page
}
\description{
This function creates a shiny dashboard page for using in Shiny app.
}
\examples{
library(shiny)
library(shinydashboard)
body <- dashboardBody()
if (interactive()) {
 # Basic dashboard page
 shinyApp(
   ui = page("Example", body),
   server = function(input, output) { }
 )
}


}
