% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_highchart.R
\name{plot_highchart}
\alias{plot_highchart}
\title{Highchart plot}
\usage{
plot_highchart(df, x, y, group, ylabs, xlabs, title)
}
\arguments{
\item{df}{The data used in producing highchart plot}

\item{x}{A date vector}

\item{y}{A character vector}

\item{group}{A discrete variable}

\item{ylabs}{The label name of yAxis}

\item{xlabs}{The label name of xAxis}

\item{title}{The title of plot}
}
\value{
Highchart plot
}
\description{
This function creates a interactive highchart plot,
which can be rendered by Shiny app.
}
\examples{
library(magrittr)
library(coronavirus)
data("coronavirus")
data <- coronavirus \%>\%
        dplyr::filter(date >= '2020-10-10') \%>\%
        dplyr::filter(country \%in\% c("Afghanistan", "Liberia", "Austria")) \%>\%
        dplyr::filter(type == "confirmed")

plot_highchart(df = data, x = date, y = cases, group = country, ylabs = 'Confirmed cases',
               xlabs = 'Date', title = 'Highchart plot for COVID-19')


}
