% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_linechart.R
\name{plot_linechart}
\alias{plot_linechart}
\title{Linechart}
\usage{
plot_linechart(df, var, date, y1, y2, y3)
}
\arguments{
\item{df}{The data used in producing line chart.}

\item{var}{A character vector.}

\item{date}{A date vector.}

\item{y1}{The first y variable.}

\item{y2}{The second y variable.}

\item{y3}{The third y variable.}
}
\value{
Linechart
}
\description{
This function creates a line chart.
}
\examples{
library(magrittr)
library(coronavirus)
library(ggplot2)

data("coronavirus")
data <- coronavirus \%>\%
 dplyr::filter(date >= '2020-09-10' & date <= '2020-10-10') \%>\%
 dplyr::filter(country \%in\% c("US", "India", "Austria")) \%>\%
 tidyr::pivot_wider(names_from = type, values_from = cases)

 plot_linechart(data, "US", date, confirmed, recovered, death)



}
